Вводит модули в проект
