+++ js/utils.js
@@ -1,0 +1,10 @@
+const getRandomInteger = (min, max) => {
+  const rand = min + Math.random() * (max + 1 - min);
+  return Math.floor(rand);
+};
+getRandomInteger(0, 20);
+
+const getRandomElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
+
+export {getRandomInteger};
+export {getRandomElement};
